+++ src/main/java/org/bigtows/window/ui/notetree/NoteTree.java
@@ -4,20 +4,18 @@ import com.intellij.ui.treeStructure.Tree;
 import org.bigtows.window.ui.notetree.listener.NoteTreeChangeListener;
 import org.bigtows.window.ui.notetree.listener.NoteTreeNeedRefreshModelListener;
 import org.bigtows.window.ui.notetree.tree.NoteTreeSelection;
-import org.bigtows.window.ui.notetree.tree.editor.PinNoteTreeCellEditor;
+import org.bigtows.window.ui.notetree.tree.PinNoteTreeCellEditor;
 import org.bigtows.window.ui.notetree.tree.entity.Note;
 import org.bigtows.window.ui.notetree.tree.entity.Task;
 import org.bigtows.window.ui.notetree.tree.node.AbstractTaskTreeNode;
 import org.bigtows.window.ui.notetree.tree.node.NoteTreeNode;
 import org.bigtows.window.ui.notetree.tree.node.TaskTreeNode;
-import org.bigtows.window.ui.notetree.tree.render.PinNoteTreeCellRender;
+import org.bigtows.window.ui.notetree.tree.PinNoteTreeCellRender;
 import org.bigtows.window.ui.notetree.utils.ExpandTreeUtils;
 
 import javax.swing.*;
-import javax.swing.event.TreeSelectionListener;
 import javax.swing.tree.*;
 import java.awt.*;
-import java.beans.PropertyChangeListener;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
